+++ solr/__init__.py
@@ -6,3 +6,4 @@ from .response import Response, Results
 from .parsers import parse_json_response, parse_query_response
 from .utils import read_response, check_response_status
 from .schema import SchemaAPI
+from .mlt import MoreLikeThis